requirements.txt
streamlit>=1.29.0
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.8.1
pandas>=2.1.4
numpy>=1.26.2
sqlparse>=0.4.4
python-dotenv>=1.0.0
pydantic>=2.5.3
ruff>=0.1.9
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.1
mypy>=1.8.0
sqlglot>=19.9.0 